refactor: split feishu plugin into platforms directory structure
- Create src/im_adapter/platforms/{mod,feishu/mod,adapter,renderer,cleaner}.rs
- Extract FeishuAdapter + token management into adapter.rs
- Extract render logic (build_card, build_text, dispatch_blocks) into renderer.rs
- Extract @ mention cleaning into cleaner.rs
- Keep FeishuPlugin + IMPlugin impl in mod.rs
- Update src/im/mod.rs to re-export from im_adapter::platforms::feishu
- Update src/im_adapter/mod.rs with pub mod platforms
- Update import paths in daemon/mod.rs, tests, plugin_tests
- Fix clippy single_match warning in plugin.rs
- Run cargo fmt for formatting compliance

Source: user
Type: refactor
