+++ src/im_adapter/platforms/mod.rs
@@ -1,0 +1,6 @@
+//! Platform-specific IM plugins.
+//!
+//! Each sub-module implements the [`IMPlugin`](super::plugin::IMPlugin) trait
+//! for a messaging platform.
+
+pub mod feishu;
