+++ tests/gateway_send_outbound_basic.rs
@@ -96,6 +96,9 @@ impl IMPlugin for TrackingPlugin {
             peer_id: "chat_1".to_string(),
             content: "hi".to_string(),
             timestamp: 0,
+            message_type: "text".to_string(),
+            media_refs: vec![],
+            quoted_message: None,
             thread_id: None,
             account_id: None,
             card_action: None,
